+++ src/app/ld/ImageFeatureSection.tsx
@@ -24,10 +24,8 @@ export interface ImageFeatureSectionProps {
     reverseLayout?: boolean
     iconColor?: string
     iconBackgroundColor?: string
-    // CTA Button Props
     ctaText?: string
     ctaHref?: string
-    ctaOnClick?: () => void
     ctaButtonColor?: string
     ctaButtonTextColor?: string
     ctaVariant?: 'contained' | 'outlined'
@@ -50,7 +48,6 @@ const ImageFeatureSection: React.FC<ImageFeatureSectionProps> = ({
     reverseLayout = false,
     ctaText,
     ctaHref,
-    ctaOnClick,
     ctaButtonColor = '#4a7c2f',
     ctaButtonTextColor = '#ffffff',
     ctaVariant = 'contained',
@@ -149,7 +146,7 @@ const ImageFeatureSection: React.FC<ImageFeatureSectionProps> = ({
     )
 
     const renderCTA = () => {
-        if (!ctaText) return null
+        if (!ctaText || !ctaHref) return null
 
         const buttonStyles = {
             px: { xs: 4, md: 5 },
@@ -182,36 +179,18 @@ const ImageFeatureSection: React.FC<ImageFeatureSectionProps> = ({
             }),
         }
 
-        if (ctaHref) {
-            return (
-                <Box sx={{ display: 'flex', justifyContent: 'center', mt: { xs: 5, md: 6 } }}>
-                    <Button
-                        component={Link}
-                        href={ctaHref}
-                        variant={ctaVariant}
-                        sx={buttonStyles}
-                    >
-                        {ctaText}
-                    </Button>
-                </Box>
-            )
-        }
-
-        if (ctaOnClick) {
-            return (
-                <Box sx={{ display: 'flex', justifyContent: 'center', mt: { xs: 5, md: 6 } }}>
-                    <Button
-                        onClick={ctaOnClick}
-                        variant={ctaVariant}
-                        sx={buttonStyles}
-                    >
-                        {ctaText}
-                    </Button>
-                </Box>
-            )
-        }
-
-        return null
+        return (
+            <Box sx={{ display: 'flex', justifyContent: 'center', mt: { xs: 5, md: 6 } }}>
+                <Button
+                    component={Link}
+                    href={ctaHref}
+                    variant={ctaVariant}
+                    sx={buttonStyles}
+                >
+                    {ctaText}
+                </Button>
+            </Box>
+        )
     }
 
     return (
