+++ src/app/ld/group-classes-indiranagar/page.tsx
@@ -1,28 +1,32 @@
-'use client'
-
 import { Box } from '@mui/material'
-import FAQCTASection from '@/app/ld/FAQCTASection'
 import GroupsIcon from '@mui/icons-material/Groups'
 import SchoolIcon from '@mui/icons-material/School'
-import FeatureSection from '@/app/ld/FeatureSection'
 import WbSunnyIcon from '@mui/icons-material/WbSunny'
-import YogaProgramHero from '@/app/ld/YogaProgramHero'
 import ScheduleIcon from '@mui/icons-material/Schedule'
-import ComparisonSection from '@/app/ld/ComparisonSection'
 import LocationOnIcon from '@mui/icons-material/LocationOn'
-const BackgroundHeroImage = '/images/anatomy.jpg'
 import PsychologyIcon from '@mui/icons-material/Psychology'
 import WorkOutlineIcon from '@mui/icons-material/WorkOutlineOutlined'
-import ImageFeatureSection from '@/app/ld/ImageFeatureSection'
-import ClassOverviewSection from '@/app/ld/ClassOverviewSection'
 import CalendarMonthIcon from '@mui/icons-material/CalendarMonth'
 import DirectionsRunIcon from '@mui/icons-material/DirectionsRun'
 import ThumbUpOffAltIcon from '@mui/icons-material/ThumbUpOffAlt'
-import TestimonialCTASection from '@/app/ld/TestimonialCTASection'
 import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder'
-const GirlMeditattngImage = '/images/mental-health/2.png'
 import SelfImprovementIcon from '@mui/icons-material/SelfImprovement'
 import PlayCircleOutlineIcon from '@mui/icons-material/PlayCircleOutlined'
+import dynamic from 'next/dynamic'
+
+const YogaProgramHero = dynamic(() => import('@/app/ld/YogaProgramHero'), { ssr: false })
+const FAQCTASection = dynamic(() => import('@/app/ld/FAQCTASection'), { ssr: false })
+import FeatureSection from '@/app/ld/FeatureSection'
+import ImageFeatureSection from '@/app/ld/ImageFeatureSection'
+import ComparisonSection from '@/app/ld/ComparisonSection'
+import ClassOverviewSection from '@/app/ld/ClassOverviewSection'
+import TestimonialCTASection from '@/app/ld/TestimonialCTASection'
+
+const BackgroundHeroImage = '/images/anatomy.jpg'
+const GirlMeditattngImage = '/images/mental-health/2.png'
+
+const WHATSAPP_URL = 'https://wa.me/+919611771434'
+const PHONE_URL = 'tel:+919611771434'
 
 export default function LandingPage() {
     return (
@@ -228,16 +232,9 @@ export default function LandingPage() {
                 ]}
                 subtext="Start Your Yoga Practice Near Indiranagar — Without Overthinking It"
                 primaryCtaText="Book a Free Trial Class"
+                primaryCtaHref={PHONE_URL}
                 secondaryCtaText="Call / WhatsApp Us for batch timings and directions"
-                onSecondaryCtaClick={() => {
-                    const phoneNumber = '+919611771434'
-                    const whatsappUrl = `https://wa.me/${phoneNumber}`
-                    window.open(whatsappUrl, '_blank')
-                }}
-                onPrimaryCtaClick={() => {
-                    const phoneNumber = '+919611771434'
-                    window.open(`tel:${phoneNumber}`, '_self')
-                }}
+                secondaryCtaHref={WHATSAPP_URL}
             />
         </Box>
     )
